Commit #2
Commit #2 starts just by adding a new string to welcome the users. In the while True loop function, I wanted this program to give however many passwords the user wants with the 'num_passwords'. Within the loop function, the 'try' will break any attempt to turn into an integer. It will also print the statement "Please enter a valid integer," if the user does not give a valid integer.