Commit #5
This commit starts the 'num_passwords.' The 'i' is a variable used as a placeholder as it won't be used in the loop. It then calls the 'generate_password' function with the two arguments. After that, I made a print statement that gives the password(s) from the user's inputs.