Commit #6
In this last commit, I called everything with main(). However, the 'if__name__ == "__main__": ' is something I found on StackOverflow that allows the program to check if the current program is being run as the main program. The reason for the main() function to be inside the 'if__name__ == "__main__": ' is that it makes sure that it is only running the program directly.